Refactor apply command modules
